+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.11.6] - 2019-10-07
+## Added
+- Add Functions `doesObjectExist` in `S3Service`
+- Add Exception in s3Service
+
 ## [0.11.5] - 2019-10-01
 ## Added
 - Add Functions `createPresignedUrl` in `S3Service`